docker configurato

+++ database/migrations/2025_08_17_150955_create_ombrellones_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('ombrellones', function (Blueprint $table) {
             $table->id();
-            $table->unsignedSmallInteger('numero')->unique();
+            $table->integer('numero')->unique();
             $table->boolean('occupato')->default(false);
             $table->timestamps();
         });
